refactor(core): relocate condition_eval ui->core (ADR-0038 Phase 2b)
Kills the last render->ui back-edge (render/dispatch.py imported
ui.utils.condition_eval for nav visibility). condition_eval is pure
(datetime + core.comparison), so core/ is its natural home — importable
downward by render, ui, and back. Updates all importers.

render/ is now fully acyclic: zero back/ and zero ui/ imports.
Full non-e2e suite: 18047 passed.
