+++ src/dazzle/render/dispatch.py
@@ -208,7 +208,7 @@ def _sidebar_from_nav_model(model: Any, ctx: PageContext) -> Sidebar:
     # context once: strip the ``role_`` prefix from roles (matching
     # page_routes' role_ctx) and pass per-tenant config so ``tenant_config.<key>``
     # references resolve. Visibility only — route access (RBAC) is unchanged.
-    from dazzle.ui.utils.condition_eval import evaluate_condition
+    from dazzle.core.condition_eval import evaluate_condition
 
     eval_ctx = {
         "user_roles": [r.removeprefix("role_") for r in (getattr(ctx, "user_roles", None) or [])],
